Split up vertex domain
