+++ pyday_night_funkin/core/graphics/shared.py
@@ -4,8 +4,12 @@ and their sizes around.
 """
 
 import ctypes
+import re
 
 from pyglet.gl import gl
+
+
+RE_VERTEX_FORMAT = re.compile("(.*)(\d)(.)(n?)/(static|dynamic|stream)")
 
 C_TYPE_MAP = {
 	gl.GL_UNSIGNED_BYTE: ctypes.c_ubyte,
